Delta versioon 3.6.22.18
Selle Delta versiooniga tulevad kaasa järgmised parandused:
* T024-320 Klassifikaatori ja klassifikaatori vaikeväärtuse kuvamine andmeväljade juures
* T024-496 Migreerimisel osade dokumentide import ebaõnnestub
* T024-719 Kasutaja ei saa lemmikute kausta midagi lisada, sest nimekiri kuvatakse lemmikute lisamise akna taha
* T024-779 Kontekstitundlik abiinfo ei tööta
* T024-845 Suurtele gruppidele saadetav töövoo käivitamine ei õnnestu, üliaeglane kuni proxy error
* T024-853 Kooskõlastaja ei saa salvestada ekraanivormile kaasa lisatud faile laiendiga .doc, docx muudatusi
* T024-904 Tähtaja pikendamise tööülesannetes otsingu väljundis tühjad veerud (struktuuriüksus, ametinimetus)
* T024-909 Rakenduse logis ei logita faili avamist
* T024-915 Dokumentide ümberliigitamisel Mozilla Firefoxiga kuvatakse teemade lahter ümberliigitamise lahtri taga ning teemasid ei ole võimalik valida
* T024-931 Süsteemsetes mallides ei asendu valemid andmetega
* T024-939 Õiguste haldus dokumentide, sarjade ja toimikute juures väga aeglane
* T024-1053 Tühjas Deltas dokumendiliikide seadistamisel algselt olemasolevatel dokumendiliikidel identifikaator väli on tühi
* T024-1058 Suletud sarjade taasavamisel tuleb viga: "$$series_open_error_inClosedFunction[]$$"
* T024-1073 Osade PP failide migreerimisel tekib viga: [Assertion failed] - this expression must be true
* Postipoisi impordil dokumendi ja sarja logi asemel rakenduse logi kirjete lisamine
* ADR'i lisakontrollide lisamine
* Klassifikaatorite importimine

Süsteemiadministraatorile: Uue versiooni paigaldamisel käivituvad järgmised mahukamad tööd:
* T024-904 jaoks: Kõikide dueDateExtensionTask'ide puhul teatud väljade väärtustamine (dueDateExtensionTaskUpdater)
* T024-845 jaoks: Kõikide tööülesannete alt failide paigutamine töövoogude alla (moveTaskFileToChildAssoc1)
* T024-845 jaoks: Kõikide tööülesannete andmete lugemine Alfresco repost ja andmete täiendamine delta_task tabelis ja sellega seotud tabelites (taskAssociatedDataTableInsertBootstrap1)
Kõikide tööde korrektsuseks peavad enne uue versiooni esmakordset käivitamist olema lucene indeksites andmed korrektses seisus.
Pärast nimetatud tööde lõppemist ja Delta edukat käivitamist ei kasutata enam tööülesannete andmeid Alfresco repost (ega lucene indeksitest). Kõikide tööülesannete kustutamine Alfresco repost (ja lucene indeksitest) ei käivitu automaatselt, kuna see võib kaua aega võtta. Selle käivitamiseks on Administraator -> Node’ide sirvija all nupp “Käivita tööülesannete kustutamise skript”, mis võimaldab antud protsessi käivitada/peatada/jätkata. Palun antud protsess edukalt lõpuni teha töövälisel ajal; seda on võimalik teha osade kaupa.

Alates sellest Delta versioonist on default organization seadetest eemaldatud PPA spetsiifilised seaded. Vaikimisi rakenduvad järgnevad seaded, mis sobivad vaikimisi teistele asutustele:
   archivals-additional=
   imap-folders.incomingLetters.subfolderType=fixed
   imap-folders.attachments.subfolderType=fixed
   imap-foders.outgoingLetters.subfolderType=fixed
   imap-folders.sendingFailed.subfolderType=fixed
Selleks, et PPA Deltas püsiksid PPA spetsiifilised seaded, on kaks võimalust:
 a) lisada PPA Delta alfresco-global.properties faili järgnevad read (esimene rida peaks live'is juba olemas olema, aga teistsuguse väärtusega, seega säilitada alfresco-global.properties failis juba olemasolev rida, kuna see on olnud kehtiv seade!):
      archivals-additional=PPA/PPA;PohjaPref/P\u00F5hja PREF;LounaPref/L\u00F5una PREF;IdaPref/Ida PREF;LaanePref/L\u00E4\u00E4ne PREF
      imap-folders.incomingLetters.subfolderType=fixed;PPA;P\u00F5hi;L\u00F5una;Ida;L\u00E4\u00E4s
      imap-folders.attachments.subfolderType=fixed;PPA;P\u00F5hi;L\u00F5una;Ida;L\u00E4\u00E4s
      imap-foders.outgoingLetters.subfolderType=fixed;PPA;P\u00F5hi;L\u00F5una;Ida;L\u00E4\u00E4s
      imap-folders.sendingFailed.subfolderType=fixed;PPA;P\u00F5hi;L\u00F5una;Ida;L\u00E4\u00E4s
 b) kasutada war faili ehitamisel conf.organization.name=ppa (kas conf-name.properties failis või andes Ant’ile -D argumendi abil)


git-svn-id: https://svn.smit/tarkvara/edhs/branches/nortal@123941 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/common/source/java/org/alfresco/repo/node/db/DbNodeServiceImpl.java
@@ -87,6 +87,8 @@ import ee.webmedia.alfresco.common.web.BeanHelper;
 import ee.webmedia.alfresco.document.model.DocumentCommonModel;
 import ee.webmedia.alfresco.template.model.DocumentTemplateModel;
 import ee.webmedia.alfresco.utils.MessageUtil;
+import ee.webmedia.alfresco.workflow.model.WorkflowCommonModel;
+import ee.webmedia.alfresco.workflow.model.WorkflowSpecificModel;
 
 /**
  * Node service using database persistence layer to fulfill functionality
@@ -801,7 +803,9 @@ public class DbNodeServiceImpl extends AbstractNodeServiceImpl
     }
     
     private void deleteChildDataFromDb(NodeRef nodeRef, QName nodeTypeQName) {
-        BeanHelper.getWorkflowDbService().deleteTasksCascading(nodeRef, nodeTypeQName);
+        if (!dictionaryService.isSubClass(getType(nodeRef), WorkflowCommonModel.Types.TASK)) {
+            BeanHelper.getWorkflowDbService().deleteTasksCascading(nodeRef, nodeTypeQName);
+        }
     }
     
     private void deletePrimaryChildren(Pair<Long, NodeRef> nodePair, boolean cascade)
@@ -2300,6 +2304,9 @@ public class DbNodeServiceImpl extends AbstractNodeServiceImpl
             if (cascade)
             {
                 pullNodeChildrenToSameStore(newChildNodePair, cascade, indexChildren);
+            }
+            if (dictionaryService.isSubClass(childNodeTypeQName, WorkflowCommonModel.Types.WORKFLOW)) {
+                BeanHelper.getWorkflowDbService().updateWorkflowTasksStore(childNodeRef, storeRef);
             }
         }
     }
